+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 **Added**:
 
 - **decidim-assemblies**: Add organizational chart to assemblies home. [\#4045](https://github.com/decidim/decidim/pull/4045)
+- **decidim-meetings**: Add upcoming events content block and page. [\#3987](https://github.com/decidim/decidim/pull/3987)
 
 **Changed**: